Footer
created a footer component and it was added to the layout page so that it was universal.  Spacing issue on right side of footer. Want to possibly use hover or buttons so that I could make the list more interactive. Links were added to left half of footer so that visitors can navigate into useful local websites.